+++ components/weather/index.ts
@@ -1,4 +1,5 @@
 export { CurrentWeather } from "./CurrentWeather";
 export { WeatherIcon, getWeatherIcon } from "./WeatherIcon";
 export { WeatherDayCard } from "./WeatherDayCard";
+export { WeatherHourCard } from "./WeatherHourCard";
 export { WeatherForecast } from "./WeatherForecast";